Changed fib example in assembly slides to factorial
