--- slides/code/08-assembly-64bit/test_fib.cpp.html
@@ -1,33 +1,0 @@
-<!DOCTYPE HTML PUBLIC "-//IETF//DTD HTML//EN">
-<html>
-<head>
-<meta http-equiv="Content-Type" content="text/html; charset=iso-8859-1">
-<meta name="GENERATOR" content="GNU source-highlight 3.1.8
-by Lorenzo Bettini
-http://www.lorenzobettini.it
-http://www.gnu.org/software/src-highlite">
-<title>test_fib.cpp</title>
-</head>
-<body bgcolor="white">
-<pre><tt><b><font color="#000080">#include</font></b> <font color="#FF0000">&lt;iostream&gt;</font>
-<b><font color="#0000FF">using</font></b> <b><font color="#0000FF">namespace</font></b> std<font color="#990000">;</font>
-
-<b><font color="#0000FF">extern</font></b> <font color="#FF0000">"C"</font> <font color="#009900">long</font> <b><font color="#000000">fib</font></b><font color="#990000">(</font><font color="#009900">unsigned</font> <font color="#009900">int</font> n<font color="#990000">);</font>
-
-<font color="#009900">long</font> <b><font color="#000000">fib</font></b><font color="#990000">(</font><font color="#009900">unsigned</font> <font color="#009900">int</font> n<font color="#990000">)</font> <font color="#FF0000">{</font>
-    <b><font color="#0000FF">if</font></b> <font color="#990000">((</font>n<font color="#990000">==</font><font color="#993399">0</font><font color="#990000">)</font> <font color="#990000">||</font> <font color="#990000">(</font>n<font color="#990000">==</font><font color="#993399">1</font><font color="#990000">))</font>
-        <b><font color="#0000FF">return</font></b> <font color="#993399">1</font><font color="#990000">;</font>
-    <b><font color="#0000FF">return</font></b> <b><font color="#000000">fib</font></b><font color="#990000">(</font>n<font color="#990000">-</font><font color="#993399">1</font><font color="#990000">)</font> <font color="#990000">+</font> <b><font color="#000000">fib</font></b><font color="#990000">(</font>n<font color="#990000">-</font><font color="#993399">2</font><font color="#990000">);</font>
-<font color="#FF0000">}</font>
-
-<font color="#009900">int</font> <b><font color="#000000">main</font></b><font color="#990000">()</font> <font color="#FF0000">{</font>
-    <font color="#009900">unsigned</font> <font color="#009900">int</font> theValue <font color="#990000">=</font> <font color="#993399">0</font><font color="#990000">;</font>
-    cout <font color="#990000">&lt;&lt;</font> <font color="#FF0000">"Enter value for fib(): "</font> <font color="#990000">&lt;&lt;</font> endl<font color="#990000">;</font>
-    cin <font color="#990000">&gt;&gt;</font> theValue<font color="#990000">;</font>
-    <font color="#009900">long</font> theResult <font color="#990000">=</font> <b><font color="#000000">fib</font></b><font color="#990000">(</font>theValue<font color="#990000">);</font>
-    cout <font color="#990000">&lt;&lt;</font> <font color="#FF0000">"The result is: "</font> <font color="#990000">&lt;&lt;</font> theResult <font color="#990000">&lt;&lt;</font> endl<font color="#990000">;</font>
-    <b><font color="#0000FF">return</font></b> <font color="#993399">0</font><font color="#990000">;</font>
-<font color="#FF0000">}</font>
-</tt></pre>
-</body>
-</html>
